refactor: modularize Rust logic
